Add lunch-time service and system time-entry flow test
Take the lunch times you entered and set them on the same date as clock_in. If the shift crosses midnight (clock_out date is the next day) and the lunch time is earlier than the clock‑in time, we roll lunch forward to the next day. If either lunch field is blank, that lunch value is cleared.